\name{buildReferenceData}
\alias{buildReferenceData}
\title{Building the default reference data of Tax4Fun2 for consistency}
\description{This functions tries to build the default reference data of Tax4Fun2.}
\arguments{
  \item{path_to_working_directory}{This function creates a new folder in the specified working directory ('Tax4Fun2_ReferenceData_v1.2') and starts to build the default reference data in this newly created folder. By default, the data is generated in the current working directory.}
	\item{use_force}{Overwrite existing results (default is FALSE)}
  \item{install_suggested_packages}{By default, the function tries to install ape and seqinr. Turn of by setting to FALSE.}
}
\author{Written by Bernd and Franziska Wemheuer}
\examples{}