\name{extractSSU}
\alias{extractSSU}
\title{Extracting SSU sequences from genomes}
\description{This function extracts SSU sequences (16S rRNA and 18S rRNA) from prokaryotic and eukaryotic genomes.}
\arguments{
  \item{genome_file}{The path to a single prokaryotic genomes}
  \item{genome_folder}{The path to a folder with prokaryotic genomes. One file per genome.}
  \item{path_to_reference_data}{The path to folder containg the reference data ("Tax4Fun2_ReferenceData_v1.1")}
	\item{file_extension}{The file extension of the genome file/files SSU sequences should be extracted from}
}
\author{Written by Bernd and Franziska Wemheuer}