\name{testReferenceData}
\alias{testReferenceData}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Testing the default reference data of Tax4Fun2 for consistency}
\description{This functions tests the default reference data of Tax4Fun2 for consistency. It will basically download a list of all files present in the data set and their md5sums to test for their presence and digital fingerprints.}
\arguments{
  \item{path_to_reference_data}{The path to folder containg the reference data ("Tax4Fun2_ReferenceData_v1.1")}
}
\author{Written by Bernd and Franziska Wemheuer}
\examples{}
